move serdes out of DurableHandler
